LayoutTests/fast/media/mq-inverted-colors-ua-styles-expected.txt

inverted-colors ua style test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getComputedStyle(document.querySelector('img')).filter is 'invert(1)'
PASS getComputedStyle(document.querySelector('picture')).filter is 'invert(1)'
PASS getComputedStyle(document.querySelector('picture>img')).filter is 'none'
PASS getComputedStyle(document.querySelector('video')).filter is 'invert(1)'
PASS successfullyParsed is true

TEST COMPLETE
